move QueryManager class from internal.dao.utils to internal.database
